splittedrhs stuff into an optional pbo
