+++ addons/grenadeLauncherAmmo/cfgAmmo.hpp
@@ -18,31 +18,4 @@ class CfgAmmo {
         simulation = "shotSmoke";
         deflectionSlowDown = 0.4;
     };
-    
-    class rhs_40mm_smoke_white: SmokeShell
-    {
-        simulation = "shotSmoke";
-        deflectionSlowDown = 0.4;
-    };
-
-    class SmokeShellRed;
-    class rhs_40mm_smoke_red: SmokeShellRed
-    {
-        simulation = "shotSmoke";
-        deflectionSlowDown = 0.4;
-    };
-
-    class SmokeShellGreen;
-    class rhs_40mm_smoke_green: SmokeShellGreen
-    {
-        simulation = "shotSmoke";
-        deflectionSlowDown = 0.4;
-    };
-
-    class SmokeShellYellow;
-    class rhs_40mm_smoke_yellow: SmokeShellYellow
-    {
-        simulation = "shotSmoke";
-        deflectionSlowDown = 0.4;
-    };
 };
