+++ addons/grenadeLauncherAmmoRhs/$PBOPREFIX$
@@ -1,0 +1,1 @@
+x\grad\addons\grenadeLauncherAmmo
