+++ addons/grenadeLauncherAmmoRhs/README.md
@@ -1,0 +1,8 @@
+### grenadeLauncherAmmo
+
+* Increases brightness of 40mm flares to a level where they are usable for illumination.
+* Decreases bounciness of 40mm smoke grenades
+
+#### Maintainer(s)
+
+* Fusselwurm
