+++ addons/grenadeLauncherAmmoRhs/cfgAmmo.hpp
@@ -1,0 +1,29 @@
+class CfgAmmo {
+    class SmokeShell;
+    class rhs_40mm_smoke_white: SmokeShell
+    {
+        simulation = "shotSmoke";
+        deflectionSlowDown = 0.4;
+    };
+
+    class SmokeShellRed;
+    class rhs_40mm_smoke_red: SmokeShellRed
+    {
+        simulation = "shotSmoke";
+        deflectionSlowDown = 0.4;
+    };
+
+    class SmokeShellGreen;
+    class rhs_40mm_smoke_green: SmokeShellGreen
+    {
+        simulation = "shotSmoke";
+        deflectionSlowDown = 0.4;
+    };
+
+    class SmokeShellYellow;
+    class rhs_40mm_smoke_yellow: SmokeShellYellow
+    {
+        simulation = "shotSmoke";
+        deflectionSlowDown = 0.4;
+    };
+};
