+++ addons/grenadeLauncherAmmoRhs/script_component.hpp
@@ -1,0 +1,4 @@
+#define COMPONENT grenadeLauncherAmmoRhs
+
+#include "\x\grad\addons\main\script_mod.hpp"
+#include "\x\grad\addons\main\script_macros.hpp"
